fix(build): 修复 Android Kotlin Long/Int 类型不匹配 + iOS CAFrameRateRange iOS 15+ 版本守卫
Android (CaptureGLProcessor.kt):
- pixelBytes 改为 Int（glBufferData 参数类型为 Int）
- glMapBufferRange offset/length 用 toLong() 显式转换

iOS (MetalRenderer.swift):
- preferredFrameRateRange 用 #available(iOS 15.0, *) 守卫
- 低版本回退到 preferredFramesPerSecond = 60

+++ flutter_app/android/app/src/main/kotlin/com/retrocam/app/camera/CaptureGLProcessor.kt
@@ -735,7 +735,7 @@ void main() {
             val pboArr = IntArray(1)
             GLES30.glGenBuffers(1, pboArr, 0)
             val pbo = pboArr[0]
-            val pixelBytes = width * height * 4L
+            val pixelBytes = width * height * 4
             GLES30.glBindBuffer(GLES30.GL_PIXEL_PACK_BUFFER, pbo)
             GLES30.glBufferData(GLES30.GL_PIXEL_PACK_BUFFER, pixelBytes, null, GLES30.GL_STREAM_READ)
             // 异步发起回读请求（GPU 继续执行其他工作）
@@ -747,7 +747,7 @@ void main() {
 
             // 7. Map PBO 读取像素（此时 GPU 传输已完成）
             val mappedBuf = GLES30.glMapBufferRange(
-                GLES30.GL_PIXEL_PACK_BUFFER, 0, pixelBytes,
+                GLES30.GL_PIXEL_PACK_BUFFER, 0, pixelBytes.toLong(),
                 GLES30.GL_MAP_READ_BIT
             ) as? java.nio.ByteBuffer
